requirements.txt
langchain==0.3.26
langchain-openai==0.3.28
openai==1.97.1
PyPDF2==3.0.1
python-dotenv==1.2.1
